+++ .gitignore
@@ -160,6 +160,7 @@ fabric.properties
 *.iml
 modules.xml
 .idea/misc.xml
+.idea
 *.ipr
 
 ### Java ###
