--- src/main/java/teamproject/AIPro/AiProApplication.java
@@ -1,13 +1,0 @@
-package teamproject.AIPro;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AiProApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AiProApplication.class, args);
-	}
-
-}
