--- src/main/java/teamproject/AIPro/config/GeminiRestTemplateConfig.java
@@ -1,22 +1,0 @@
-package teamproject.AIPro.config;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-@RequiredArgsConstructor
-public class GeminiRestTemplateConfig {
-    @Bean
-    @Qualifier("geminiRestTemplate")
-    public RestTemplate geminiRestTemplate() {
-        RestTemplate restTemplate = new RestTemplate();
-        restTemplate.getInterceptors().add((request, body, execution) -> execution.execute(request, body));
-
-        return restTemplate;
-    }
-
-}
