--- src/main/java/teamproject/AIPro/domain/chat/dto/request/ChatHistoryRequest.java
@@ -1,13 +1,0 @@
-package teamproject.AIPro.domain.chat.dto.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ChatHistoryRequest {
-    private String userId;
-    private String question;
-    private String response;
-
-}
