--- src/main/java/teamproject/AIPro/domain/chat/dto/request/ChatRequest.java
@@ -1,11 +1,0 @@
-package teamproject.AIPro.domain.chat.dto.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ChatRequest {
-    private String userId;
-    private String question;
-}
