--- src/main/java/teamproject/AIPro/domain/chat/dto/response/ChatResponse.java
@@ -1,10 +1,0 @@
-package teamproject.AIPro.domain.chat.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ChatResponse {
-    private String message;
-}
