--- src/main/java/teamproject/AIPro/domain/chat/repository/ChatHistoryRepository.java
@@ -1,12 +1,0 @@
-package teamproject.AIPro.domain.chat.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import teamproject.AIPro.domain.chat.entity.ChatHistory;
-
-import java.util.List;
-
-@Repository
-public interface ChatHistoryRepository extends JpaRepository<ChatHistory, Long> {
-    List<ChatHistory> findByUserId(String userId);
-}
