--- src/main/java/teamproject/AIPro/domain/member/dto/request/LoginRequest.java
@@ -1,11 +1,0 @@
-package teamproject.AIPro.domain.member.dto.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginRequest {
-    private String email;
-    private String password;
-}
