--- src/main/java/teamproject/AIPro/domain/member/dto/request/SignupRequest.java
@@ -1,12 +1,0 @@
-package teamproject.AIPro.domain.member.dto.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class SignupRequest {
-    private String email;
-    private String password;
-    private String username;
-}
