--- src/main/java/teamproject/AIPro/domain/member/dto/response/MemberResponse.java
@@ -1,13 +1,0 @@
-package teamproject.AIPro.domain.member.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class MemberResponse {
-    
-    private Long id;
-    private String email;
-    private String username;
-}
