--- src/main/java/teamproject/AIPro/domain/member/entity/Member.java
@@ -1,31 +1,0 @@
-package teamproject.AIPro.domain.member.entity;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import jakarta.persistence.*;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String username;
-
-    public Member(String email, String password, String username) {
-        this.email = email;
-        this.password = password;
-        this.username = username;
-    }
-}
