--- src/main/java/teamproject/AIPro/domain/member/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package teamproject.AIPro.domain.member.repository;
-import teamproject.AIPro.domain.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
-}
