--- src/main/java/teamproject/AIPro/domain/role/dto/request/RoleRequest.java
@@ -1,10 +1,0 @@
-package teamproject.AIPro.domain.role.dto.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class RoleRequest {
-    private String role;
-}
