--- src/main/java/teamproject/AIPro/domain/role/dto/response/RoleResponse.java
@@ -1,12 +1,0 @@
-package teamproject.AIPro.domain.role.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class RoleResponse {
-    private String role;
-}
