+++ src/main/java/teamproject/aipro/AiProApplication.java
@@ -1,0 +1,13 @@
+package teamproject.aipro;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AiProApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AiProApplication.class, args);
+	}
+
+}
