+++ src/main/java/teamproject/aipro/domain/chat/dto/request/ChatHistoryRequest.java
@@ -1,0 +1,13 @@
+package teamproject.aipro.domain.chat.dto.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ChatHistoryRequest {
+	private String userId;
+	private String question;
+	private String response;
+
+}
