+++ src/main/java/teamproject/aipro/domain/chat/dto/response/ChatResponse.java
@@ -1,0 +1,10 @@
+package teamproject.aipro.domain.chat.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ChatResponse {
+	private String message;
+}
