+++ src/main/java/teamproject/aipro/domain/chat/repository/ChatHistoryRepository.java
@@ -1,0 +1,13 @@
+package teamproject.aipro.domain.chat.repository;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import teamproject.aipro.domain.chat.entity.ChatHistory;
+
+@Repository
+public interface ChatHistoryRepository extends JpaRepository<ChatHistory, Long> {
+	List<ChatHistory> findByUserId(String userId);
+}
