+++ src/main/java/teamproject/aipro/domain/member/dto/request/LoginRequest.java
@@ -1,0 +1,11 @@
+package teamproject.aipro.domain.member.dto.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequest {
+	private String email;
+	private String password;
+}
