+++ src/main/java/teamproject/aipro/domain/member/dto/request/SignupRequest.java
@@ -1,0 +1,12 @@
+package teamproject.aipro.domain.member.dto.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class SignupRequest {
+	private String email;
+	private String password;
+	private String username;
+}
