+++ src/main/java/teamproject/aipro/domain/member/dto/response/MemberResponse.java
@@ -1,0 +1,13 @@
+package teamproject.aipro.domain.member.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class MemberResponse {
+
+	private Long id;
+	private String email;
+	private String username;
+}
