+++ src/main/java/teamproject/aipro/domain/member/repository/MemberRepository.java
@@ -1,0 +1,11 @@
+package teamproject.aipro.domain.member.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import teamproject.aipro.domain.member.entity.Member;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+	Optional<Member> findByEmail(String email);
+}
