+++ src/main/java/teamproject/aipro/domain/role/dto/request/RoleRequest.java
@@ -1,0 +1,10 @@
+package teamproject.aipro.domain.role.dto.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class RoleRequest {
+	private String role;
+}
