+++ src/main/java/teamproject/aipro/domain/role/dto/response/RoleResponse.java
@@ -1,0 +1,12 @@
+package teamproject.aipro.domain.role.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class RoleResponse {
+	private String role;
+}
